make industry averages

+++ App/db/migrate/20150807230538_create_industries.rb
@@ -2,15 +2,15 @@ class CreateIndustries < ActiveRecord::Migration
   def change
     create_table :industries do |t|
       t.string :name
-      t.integer :eps
-      t.integer :pe
-      t.integer :pbook
-      t.integer :psales
-      t.integer :markcap
-      t.integer :peg
-      t.integer :book_value
-      t.integer :shares
-      t.integer :graham_number
+      t.float :eps
+      t.float :pe
+      t.float :pbook
+      t.float :psales
+      t.float :markcap
+      t.float :peg
+      t.float :book_value
+      t.float :shares
+      t.float :graham_number
 
       t.timestamps null: false
     end
